Facebook style

+++ src/route.rs
@@ -17,7 +17,7 @@ macro_rules! redirect {
 pub(crate) use redirect;
 
 pub fn save_file(target_dir: &PathBuf, temp_file: Option<TempFile>, file_name: &str) -> Result<(String, String), ()> {
-    let mut md5sum = String::new();
+    let md5sum;
 
     if let Some(f) = temp_file {
         if !target_dir.exists() {
@@ -26,7 +26,6 @@ pub fn save_file(target_dir: &PathBuf, temp_file: Option<TempFile>, file_name: &
                 return Err(());
             }
         }
-
 
         let mut hasher = Md5::new();
         let temp_file_name = f.file_name.unwrap_or_default();
@@ -48,20 +47,24 @@ pub fn save_file(target_dir: &PathBuf, temp_file: Option<TempFile>, file_name: &
         hasher.update(&buf);
         md5sum = format!("{:x}", hasher.finalize());
 
-        let mut saved_name = if file_name.is_empty() {
+        let saved_name = if file_name.is_empty() {
             format!("{}.{}", md5sum, extension)
         } else {
             String::from(file_name)
         };
 
         let path = target_dir.join(saved_name.as_str());
-        if let Err(e) = f.file.persist(&path) {
-            log::warn!("Failed to save file to {:?}: {}", &path.to_str(), e);
-            log::warn!("Try copying");
-            if let Err(e) = fs::copy(e.file.path(), &path) {
-                log::error!("Failed to copy file: {}", e);
-                return Err(());
+        if !path.exists() {
+            if let Err(e) = f.file.persist(&path) {
+                log::warn!("Failed to save file to {:?}: {}", &path.to_str(), e);
+                log::warn!("Try copying");
+                if let Err(e) = fs::copy(e.file.path(), &path) {
+                    log::error!("Failed to copy file: {}", e);
+                    return Err(());
+                }
             }
+        } else {
+            log::info!("File {saved_name} exists");
         }
 
         return Ok((md5sum, saved_name));
@@ -69,4 +72,11 @@ pub fn save_file(target_dir: &PathBuf, temp_file: Option<TempFile>, file_name: &
     Ok((String::new(), String::new()))
 }
 
-pub fn delete_file(file_path: String) {}
+pub fn delete_file(root_dir: &PathBuf, file_path: String) {
+    if !file_path.is_empty() {
+        let real_path = root_dir.join(&file_path);
+        if let Err(e) = fs::remove_file(&real_path) {
+            log::error!("Failed to remove file {}: {}", file_path, e);
+        }
+    }
+}
